requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Vector Database and Embeddings
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# Document Processing
PyMuPDF==1.23.8
PyPDF2==3.0.1
python-docx==1.1.0

# LLM Integration
openai==1.3.7
tiktoken==0.5.2

# Data Processing
numpy==1.24.3
pandas==2.1.4

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
requests==2.31.0
tqdm==4.66.1

# Development
pytest==7.4.3
black==23.12.0
flake8==6.1.0